pyproject.toml
[project]
name = "dio-blog"
version = "0.1.0"
description = ""
authors = [
    {name = "Cassiano Shigueyuki Nishikawa",email = "csnishikawa@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi",
    "uvicorn[standard] (>=0.40.0,<0.41.0)",
    "databases[aiosqlite] (>=0.9.0,<0.10.0)",
    "pyjwt",
    "psycopg2-binary",
    "pydantic-settings"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "pytest (>=9.0.2,<10.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "pytest-mock"
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
